# api-gateway/internal/docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: E-Commerce Microservices Platform — Public API
  version: 1.0.0
  description: |
    Toàn bộ API public đi qua `api-gateway` (`:8080`). Postgres/Redis/RabbitMQ và
    các service phía sau **không** publish port ra host, nên đây là bề mặt duy
    nhất client gọi được.

    ## Response envelope
    Mọi response JSON đều dùng chung một envelope (`pkg/response`):
    ```json
    { "success": true, "message": "...", "data": {...}, "error": "...", "meta": {...} }
    ```
    `data` chỉ có khi thành công, `error` chỉ có khi lỗi, `meta` chỉ có ở API phân trang.

    ## Xác thực
    Gửi `Authorization: Bearer <access_token>` lấy từ `POST /api/v1/auth/login`.
    Route gắn khoá 🔒 yêu cầu JWT; phần mô tả ghi rõ role nào được phép.

    ## Idempotency
    `POST /api/v1/orders`, `POST /api/v1/payments` và `POST /api/v1/admin/payments/{id}/refunds`
    nhận header `Idempotency-Key`. Gửi lại cùng key sẽ trả về kết quả của lần
    xử lý đầu tiên thay vì tạo bản ghi mới.

    ## Rate limit
    Gateway giới hạn theo IP bằng Redis (60 req/phút, burst 120, chặn 2 phút khi
    vượt). Vượt ngưỡng trả `429`.
servers:
  - url: http://localhost:8080
    description: api-gateway (gọi trực tiếp, dùng khi dev/test)
  - url: http://localhost
    description: Nginx edge (đứng trước gateway)
...
security: []
paths:
...
  /api/v1/cart:
    get:
      tags: [Cart]
      summary: 🔒 Xem giỏ hàng
      description: Giỏ hàng lưu trên Redis kèm TTL, không có bảng SQL.
      operationId: getCart
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Giỏ hàng hiện tại.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CartEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      tags: [Cart]
      summary: 🔒 Xoá sạch giỏ hàng
      operationId: clearCart
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Giỏ hàng rỗng.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CartEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/cart/items:
    post:
      tags: [Cart]
      summary: 🔒 Thêm sản phẩm vào giỏ
      description: Giá và tên lấy từ product-service qua gRPC, không tin số client gửi lên.
      operationId: addCartItem
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddToCartRequest' }
      responses:
        '200':
          description: Giỏ hàng sau khi thêm.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CartEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/cart/merge:
    post:
      tags: [Cart]
      summary: 🔒 Hợp nhất giỏ hàng khách vãng lai sau khi đăng nhập
      operationId: mergeCart
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MergeCartRequest' }
      responses:
        '200':
          description: Giỏ hàng sau khi hợp nhất.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CartEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/cart/items/{productId}:
    parameters:
      - $ref: '#/components/parameters/ProductIdPath'
      - name: sku
        in: query
        required: false
        schema: { type: string, maxLength: 120 }
        description: >-
          Chọn đúng dòng giỏ hàng khi sản phẩm có nhiều variant. Bỏ trống là trỏ
          tới dòng không có variant của sản phẩm đó.
    put:
      tags: [Cart]
      summary: 🔒 Đổi số lượng một item
      operationId: updateCartItem
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateCartItemRequest' }
      responses:
        '200':
          description: Giỏ hàng sau khi cập nhật.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CartEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Cart]
      summary: 🔒 Bỏ một item khỏi giỏ
      operationId: removeCartItem
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Giỏ hàng sau khi xoá.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CartEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ── Catalog / Coupons (public) ──────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token lấy từ `POST /api/v1/auth/login`.
  parameters:
...
    ProductIdPath:
      name: productId
      in: path
      required: true
      schema: { type: string }
      description: ID sản phẩm.
...
  responses:
    BadRequest:
      description: Body hoặc tham số không hợp lệ.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: Thiếu token, token sai hoặc hết hạn.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    NotFound:
      description: Không tìm thấy tài nguyên.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
...
    Meta:
      type: object
      description: Metadata phân trang. Chỉ xuất hiện ở API có phân trang.
      properties:
        page: { type: integer, example: 1 }
        limit: { type: integer, example: 20 }
        total: { type: integer, format: int64, example: 137 }
        next_cursor:
          type: string
          description: Con trỏ trang kế tiếp khi dùng cursor pagination.
        has_next: { type: boolean }
    ApiEnvelope:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean, example: true }
        message: { type: string, example: request succeeded }
        data: {}
        error: { type: string }
        meta: { $ref: '#/components/schemas/Meta' }
    ErrorEnvelope:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean, example: false }
        message: { type: string, example: validation failed }
        error:
          type: string
          description: Chi tiết lỗi đọc được.
          example: email is required
...
    CartItem:
      type: object
      description: >-
        Một dòng giỏ hàng được định danh bằng cặp (`product_id`, `sku`): hai
        variant của cùng một sản phẩm là hai dòng riêng, có số lượng và giá riêng.
      properties:
        product_id: { type: string }
        sku: { type: string, description: Variant của dòng này; rỗng nếu sản phẩm không có variant. }
        variant_label: { type: string, description: Nhãn hiển thị của variant, ví dụ "Đen / size M". }
        name: { type: string }
        price: { type: number, format: double }
        quantity: { type: integer }
    Cart:
      type: object
      properties:
        user_id: { type: string }
        items: { type: array, items: { $ref: '#/components/schemas/CartItem' } }
        total: { type: number, format: double }

    # ── Orders ────────────────────────────────────────────────────────────────
...
    AddToCartRequest:
      type: object
      required: [product_id, quantity]
      properties:
        product_id: { type: string }
        sku:
          type: string
          maxLength: 120
          description: >-
            Variant muốn mua. Bắt buộc khi sản phẩm có khai báo `variants`; bỏ
            trống sẽ nhận 400 vì không xác định được size/màu nào bị trừ kho.
            Sản phẩm không có variant thì để trống.
        quantity: { type: integer, minimum: 1 }
    UpdateCartItemRequest:
      type: object
      required: [quantity]
      properties:
        quantity: { type: integer, minimum: 1 }
    MergeCartRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/AddToCartRequest' }
...
    CartEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/Cart' }
...